LPS-151265 search-experiences-test:rename test file names
https://issues.liferay.com/browse/LPS-151265
